# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "LibreOffice 文档转换 REST API",
        "title": "libreoffice-rest-api",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:30231",
    "basePath": "/",
    "paths": {
        "/api/v1/convert": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "上传文件或 URL 提交转换任务，支持同步与异步模式",
                "consumes": [
                    "multipart/form-data",
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "提交转换任务",
                "parameters": [
                    {
                        "type": "file",
                        "description": "上传文件",
                        "name": "file",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "输出格式（常用：pdf/html/png/txt/odt/doc/docx/rtf/epub/xls/xlsx/ods/csv/ppt/pptx/odp/odg/svg/jpg/jpeg/webp）",
                        "name": "format",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "sync/async",
                        "name": "mode",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "URL 下载地址",
                        "name": "url",
                        "in": "formData"
                    },
                    {
                        "description": "JSON 请求体",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/http.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/files/{id}/download": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "summary": "下载转换结果",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/merge": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "上传多个文件或 URL 合并为单个 PDF，支持同步与异步模式",
                "consumes": [
                    "multipart/form-data",
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "提交合并任务",
                "parameters": [
                    {
                        "type": "file",
                        "description": "上传文件（可多次传入）",
                        "name": "files",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "输出格式（仅支持 pdf）",
                        "name": "format",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "sync/async",
                        "name": "mode",
                        "in": "formData"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "URL 列表",
                        "name": "urls",
                        "in": "formData"
                    },
                    {
                        "description": "JSON 请求体",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/http.MergeRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/http.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/tasks/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "查询任务状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.APIResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "健康检查",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/http.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/http.ErrorDetail"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "http.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "http.MergeRequest": {
            "type": "object",
            "properties": {
                "format": {
                    "type": "string"
                },
                "mode": {
                    "type": "string"
                },
                "urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}